Fixed bugs in MsSql to MySql replication
Table Sales replicates just fine now

+++ Test C# Solution/Supermarket/Supermarket.ImportExel/ImportExel.cs
@@ -123,7 +123,7 @@ namespace Supermarket.ImportExel
                     SupermarketId = marketId,
                     ProductId = productId,
                     Quantity = quantity,
-                    Price = price,
+                    Price = (decimal) price,
                     SalesDate = dateReport
                 });
                 context.SaveChanges();
